Clean up -.-

+++ src/main/java/com/sjkz1/minetils/render/RainBowAxolotl.java
@@ -1,5 +1,7 @@
 package com.sjkz1.minetils.render;
 
+import java.awt.Color;
+
 import net.minecraft.client.render.OverlayTexture;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexConsumer;
@@ -10,7 +12,6 @@ import net.minecraft.client.render.entity.model.AxolotlEntityModel;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.entity.passive.AxolotlEntity;
 import net.minecraft.util.Identifier;
-import java.awt.*;
 
 public class RainBowAxolotl<T extends AxolotlEntity, M extends AxolotlEntityModel<T>> extends FeatureRenderer<T, M>
 {
@@ -23,7 +24,7 @@ public class RainBowAxolotl<T extends AxolotlEntity, M extends AxolotlEntityMode
 
     @Override
     public void render(MatrixStack matrixStack, VertexConsumerProvider vertexConsumerProvider, int i, T entity, float f, float g, float h, float j, float k, float l) {
-        float x = ((float)((entity.age + entity.age) % 90) + g) / 90.0F;
+        float x = ((entity.age + entity.age) % 90 + g) / 90.0F;
         Color color = Color.getHSBColor(x,0.9f,1);
         RenderLayer renderLayer = RenderLayer.getEntityCutoutNoCull(SKIN);
         VertexConsumer vertexConsumer = vertexConsumerProvider.getBuffer(renderLayer);
